+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/synthetic/JoinedRecordPlan.java
@@ -85,7 +85,7 @@ class JoinedRecordPlan implements SyntheticRecordFromStoredRecordPlan  {
             for (BindingPlan bindingPlan : bindingPlans) {
                 builder.setBinding(bindingPlan.name, bindingPlan.evaluate(record));
             }
-            return builder.build(context.getTypeRepository());
+            return builder.build();
         }
 
         @Override